Extract users local config
